Former-commit-id: 48e42e63c13dae0e4dc53d766e791f63b705d335

+++ ifs-data-service/src/main/java/org/innovateuk/ifs/assessment/controller/AssessmentController.java
@@ -44,7 +44,7 @@ public class AssessmentController {
     }
 
     @RequestMapping(value = "/state/{state}/competition/{competitionId}/count", method = GET)
-    public RestResult<Long> countByStateAndCompetition(
+    public RestResult<Integer> countByStateAndCompetition(
             @PathVariable("state") AssessmentStates state,
             @PathVariable("competitionId") Long competitionId) {
         return assessmentService.countByStateAndCompetition(state, competitionId).toGetResponse();
